pyproject.toml
[project]
name = "flaskproject"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.3",
    "mysql-connector-python>=9.6.0",
    "pytest>=9.0.2",
]